requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pdflatex>=0.1.0
